Move the `LZWStream` from `src/core/stream.js` and into its own file
