+++ src/core/parser.js
@@ -17,7 +17,6 @@ import {
   Ascii85Stream,
   AsciiHexStream,
   FlateStream,
-  LZWStream,
   NullStream,
   PredictorStream,
   RunLengthStream,
@@ -47,6 +46,7 @@ import { CCITTFaxStream } from "./ccitt_stream.js";
 import { Jbig2Stream } from "./jbig2_stream.js";
 import { JpegStream } from "./jpeg_stream.js";
 import { JpxStream } from "./jpx_stream.js";
+import { LZWStream } from "./lzw_stream.js";
 
 const MAX_LENGTH_TO_CACHE = 1000;
 const MAX_ADLER32_LENGTH = 5552;
